kern_proc_vmmap_out: paged interface
In particular, allow a start address that isn't 0.  This will let
us use a fixed-size buffer in userland no matter how complicated the
address space gets, which seems like a good idea.  While here, also
allow a limit of how many structures the kernel should enumerate;
this stops us from getting spurious ENOMEMs even when some structures
could be placed in the designated buffer.

This is something of a hack, using the "new" pointer field of the sysctl,
but that's probably fine, as that field is otherwise unusable, this being
a read-only sysctl.