DNM: kern_proc_vmmap_out: fabricate capability
Yikes!  Return a capability to userland so that the caller of this sysctl
can wander around inside it.

This is absolutely unsafe as written and is just for testing.  (Though it
remains fine if all one cares about is sandboxes vs. not, since sandboxes
can't make syscalls.) Going forward, we will need to guard the authority to
do this, somehow.